Add files via upload
Made the display more dynamic. If it's night time in the region, then the sun is swapped with the moon, the sky is darkened, and stars are present. If the weather is cloudy in the region then the clouds will be darker. If the weather description mentions rain then rain animations are present. If snow is mentioned then snow animations are present. Bugs may still be present.